requirements.txt

ocpp==0.14.0
websockets==10.4
#requirements.txt
aiohttp>=3.9.0
psycopg2-binary
asyncpg>=0.27.0
fastapi==0.111.0
uvicorn==0.29.0
pydantic==2.7.1
sqlalchemy
flask-sqlalchemy

Flask>=2.0.0
Flask-RESTX>=0.5.0
python-dotenv>=0.19.0
python-dateutil>=2.8.0
pyyaml>=6.0.0
gunicorn>=20.0.0